+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## [Unreleased]
 ### Added
+- Added check to highlight if fields are placed in the wrong paragraph in debian/control files. (Closes: #1)
 
 ### Changed